tunable2choice wring result

+++ tunableTwoChoice/Amortized.h
@@ -38,8 +38,8 @@ private:
 
 public:
     Amortized(long N, bool inMemory, bool overwrite);
-    void update(OP op, string keyword, long ind, bool setup);
-    vector<long> search(string keyword);
+    void update(OP op, string keyword, int ind, bool setup);
+    vector<int> search(string keyword);
     virtual ~Amortized();
     double getTotalSearchCommSize() const;
     double getTotalUpdateCommSize() const;
